+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ------------
 NEW FEATURES
 
+* new bedTracks() function generating BED6-style GRangesList for circRNA browser tracks
+
 IMPROVEMENTS AND BUG FIXES